+++ mne/channels/__init__.py
@@ -8,5 +8,3 @@ from .layout import (Layout, make_eeg_layout, make_grid_layout, read_layout,
 
 from .channels import (equalize_channels, rename_channels,
                        read_ch_connectivity)
-
-from . import utils