+++ mne/cov.py
@@ -177,7 +177,8 @@ class Covariance(dict):
 ###############################################################################
 # IO
 
-def read_cov(fname):
+@verbose
+def read_cov(fname, verbose=None):
     """Read a noise covariance from a FIF file.
 
     Parameters
@@ -185,6 +186,8 @@ def read_cov(fname):
     fname : string
         The name of file containing the covariance matrix. It should end with
         -cov.fif or -cov.fif.gz.
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see mne.verbose).
 
     Returns
     -------
@@ -198,6 +201,41 @@ def read_cov(fname):
 
 ###############################################################################
 # Estimate from data
+
+def make_ad_hoc_cov(info):
+    """Create an ad hoc noise covariance
+
+    Parameters
+    ----------
+    info : instance of mne.io.meas_info.Info
+        Measurement info.
+
+    Returns
+    -------
+    cov : instance of Covariance
+        The ad hoc diagonal noise covariance for the M/EEG data channels.
+    """
+    info = pick_info(info, pick_types(info, meg=True, eeg=True))
+
+    # Standard deviations to be used
+    grad_std = 5e-13
+    mag_std = 20e-15
+    eeg_std = 0.2e-6
+    logger.info('Using standard noise values '
+                '(MEG grad : %6.1f fT/cm MEG mag : %6.1f fT EEG : %6.1f uV)'
+                % (1e13 * grad_std, 1e15 * mag_std, 1e6 * eeg_std))
+
+    data = np.zeros(len(info['ch_names']))
+    for meg, eeg, val in zip(('grad', 'mag', False), (False, False, True),
+                             (grad_std, mag_std, eeg_std)):
+        data[pick_types(info, meg=meg, eeg=eeg)] = val * val
+    cov = Covariance(None)
+    cov.update(kind=FIFF.FIFFV_MNE_NOISE_COV, diag=True, dim=len(data),
+               names=info['ch_names'], data=data, projs=info['projs'],
+               bads=info['bads'], nfree=0, eig=None, eigvec=None,
+               info=info)
+    return cov
+
 
 def _check_n_samples(n_samples, n_chan):
     """Check to see if there are enough samples for reliable cov calc"""
@@ -824,7 +862,7 @@ def _auto_low_rank_model(data, mode, n_jobs, method_params, cv,
         if score != -np.inf:
             scores[ii] = score
 
-        if (ii >= 3 and np.all(np.diff(scores[ii-3:ii]) < 0.) and
+        if (ii >= 3 and np.all(np.diff(scores[ii - 3:ii]) < 0.) and
            stop_early is True):
             # early stop search when loglik has been going down 3 times
             logger.info('early stopping parameter search.')
@@ -872,6 +910,7 @@ if check_sklearn_version('0.15') is True:
             cov_['nfree'] = len(self.covariance_)
             cov_['bads'] = self.info['bads']
             cov_['projs'] = self.info['projs']
+            cov_['diag'] = False
             cov_ = regularize(cov_, self.info, grad=self.grad, mag=self.mag,
                               eeg=self.eeg, proj=False,
                               exclude='bads')  # ~proj == important!!
@@ -988,8 +1027,8 @@ def _unpack_epochs(epochs):
     return epochs
 
 
-@verbose
-def _get_whitener(A, pca, ch_type, rank=None, verbose=None):
+def _get_ch_whitener(A, pca, ch_type, rank):
+    """"Get whitener params for a set of channels"""
     # whitening operator
     eig, eigvec = linalg.eigh(A, overwrite_a=True)
     eigvec = eigvec.T
@@ -1035,15 +1074,12 @@ def prepare_noise_cov(noise_cov, info, ch_names, rank=None,
     """
     C_ch_idx = [noise_cov.ch_names.index(c) for c in ch_names]
     if noise_cov['diag'] is False:
-
         C = noise_cov.data[np.ix_(C_ch_idx, C_ch_idx)]
     else:
         C = np.diag(noise_cov.data[C_ch_idx])
 
     scalings_ = dict(mag=1e12, grad=1e11, eeg=1e5)
-    if scalings is None:
-        pass
-    elif isinstance(scalings, dict):
+    if isinstance(scalings, dict):
         scalings_.update(scalings)
 
     # Create the projection operator
@@ -1083,8 +1119,8 @@ def prepare_noise_cov(noise_cov, info, ch_names, rank=None,
             if len(C_meg_idx) < len(pick_meg):
                 this_info = pick_info(info, C_meg_idx)
             rank_meg = _estimate_rank_meeg_cov(C_meg, this_info, scalings_)
-        C_meg_eig, C_meg_eigvec = _get_whitener(C_meg, False, 'MEG',
-                                                rank_meg)
+        C_meg_eig, C_meg_eigvec = _get_ch_whitener(C_meg, False, 'MEG',
+                                                   rank_meg)
     if has_eeg:
         C_eeg = C[np.ix_(C_eeg_idx, C_eeg_idx)]
         this_info = pick_info(info, pick_eeg)
@@ -1092,8 +1128,8 @@ def prepare_noise_cov(noise_cov, info, ch_names, rank=None,
             if len(C_meg_idx) < len(pick_meg):
                 this_info = pick_info(info, C_eeg_idx)
             rank_eeg = _estimate_rank_meeg_cov(C_eeg, this_info, scalings_)
-        C_eeg_eig, C_eeg_eigvec = _get_whitener(C_eeg, False, 'EEG',
-                                                rank_eeg)
+        C_eeg_eig, C_eeg_eigvec = _get_ch_whitener(C_eeg, False, 'EEG',
+                                                   rank_eeg)
         if not _has_eeg_average_ref_proj(info['projs']):
             warnings.warn('No average EEG reference present in info["projs"], '
                           'covariance may be adversely affected. Consider '
@@ -1291,8 +1327,9 @@ def compute_whitener(noise_cov, info, picks=None, rank=None,
     return W, ch_names
 
 
-def whiten_evoked(evoked, noise_cov, picks, diag=False, rank=None,
-                  scalings=None):
+@verbose
+def whiten_evoked(evoked, noise_cov, picks=None, diag=False, rank=None,
+                  scalings=None, verbose=None):
     """Whiten evoked data using given noise covariance
 
     Parameters
@@ -1301,8 +1338,9 @@ def whiten_evoked(evoked, noise_cov, picks, diag=False, rank=None,
         The evoked data
     noise_cov : instance of Covariance
         The noise covariance
-    picks : array-like of int
-        The channel indices to whiten
+    picks : array-like of int | None
+        The channel indices to whiten. Can be None to whiten MEG and EEG
+        data.
     diag : bool
         If True, whiten using only the diagonal of the covariance.
     rank : None | int | dict
@@ -1318,29 +1356,39 @@ def whiten_evoked(evoked, noise_cov, picks, diag=False, rank=None,
 
             dict(mag=1e15, grad=1e13, eeg=1e6)
 
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see mne.verbose).
+
     Returns
     -------
     evoked_white : instance of Evoked
         The whitened evoked data.
     """
-    ch_names = [evoked.ch_names[k] for k in picks]
     evoked = cp.deepcopy(evoked)
+    if picks is None:
+        picks = pick_types(evoked.info, meg=True, eeg=True)
+    W = _get_whitener_data(evoked.info, noise_cov, picks,
+                           diag, rank, scalings, evoked.nave)
+    evoked.data[picks] = np.sqrt(evoked.nave) * np.dot(W, evoked.data[picks])
+    return evoked
 
+
+@verbose
+def _get_whitener_data(info, noise_cov, picks, diag=False, rank=None,
+                       scalings=None, verbose=None):
+    """Get whitening matrix for a set of data"""
+    ch_names = [info['ch_names'][k] for k in picks]
     noise_cov = pick_channels_cov(noise_cov, include=ch_names, exclude=[])
     if diag:
         noise_cov = cp.deepcopy(noise_cov)
         noise_cov['data'] = np.diag(np.diag(noise_cov['data']))
 
     scalings_ = dict(mag=1e12, grad=1e11, eeg=1e5)
-    if scalings is None:
-        pass
-    elif isinstance(scalings, dict):
+    if isinstance(scalings, dict):
         scalings_.update(scalings)
 
-    W, _ = compute_whitener(noise_cov, evoked.info, rank=rank,
-                            scalings=scalings)
-    evoked.data[picks] = np.sqrt(evoked.nave) * np.dot(W, evoked.data[picks])
-    return evoked
+    W = compute_whitener(noise_cov, info, rank=rank, scalings=scalings_)[0]
+    return W
 
 
 @verbose