+++ doc/source/whats_new.rst
@@ -9,6 +9,10 @@ Current
 Changelog
 ~~~~~~~~~
 
+   - Add support for single dipole fitting by `Eric Larson`_
+
+   - Add support for spherical models in forward calculations by `Eric Larson`_
+
    - Add support for mayavi figures in `add_section` method in Report by `Mainak Jas`_
 
    - Add extract volumes of interest from freesurfer segmentation and setup as volume source space by `Alan Leggitt`_